+++ Assets/Framework/GameActions/[EffectDurationTick].cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0dc2795570433f547af1ae2e86602e8a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
